104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,12 +9,12 @@
 
 int main(void)
 {
-	int num1 = 1;
-	int num2 = 2;
-	int num3;
+	unsigned long int num1 = 1;
+	unsigned long int num2 = 2;
+	unsigned long int num3;
 	int i;
 
-	printf("%d, %d, ", num1, num2);
+	printf("%lu, %lu, ", num1, num2);
 
 	for (i = 0; i < 98; i++)
 	{
@@ -22,7 +22,7 @@ int main(void)
 		num1 = num2;
 		num2 = num3;
 
-		printf("%d, ", num3);
+		printf("%lu, ", num3);
 	}
 
 	return (0);
